requirements.txt
adafruit-io==2.5.0
pimoroni-sgp30==0.0.2
pyserial
